requirements.txt
pvporcupine==3.0.0
pvleopard==2.0.0
openai==1.0.0
requests==2.31.0
pydub==0.25.1
playsound==1.3.0
asyncio-mqtt==0.16.1
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
sounddevice
soundfile
pydub
numpy
